+++ components/Proposal/ProposalList.tsx
@@ -1,16 +1,14 @@
-import { useQuery } from '@apollo/client';
 import Card from '@material-ui/core/Card';
 import Grid from '@material-ui/core/Grid';
 import { createStyles, makeStyles } from '@material-ui/core/styles';
 import Typography from '@material-ui/core/Typography';
 import React from 'react';
 
-import { GET_PROPOSALS } from '../Query/Proposal';
 import Chips from '../Utils/Chips';
 import ComponentLoader from '../Utils/ComponentLoader';
-import ErrorMessage from '../Utils/ErrorMessage';
 import MiddleEllipsis from '../Utils/MiddleEllipsis';
 import NavLink from '../Utils/NavLink';
+import NotAvailable from '../Utils/NotAvailable';
 
 const useStyles = makeStyles(() =>
     createStyles({
@@ -65,140 +63,82 @@ const useStyles = makeStyles(() =>
     })
 );
 
-type ProposalListProps = { title: string[] };
+type ProposalListProps = { title: any };
 
 const ProposalList = ({ title }: ProposalListProps): JSX.Element => {
     const classes = useStyles();
-    const page = process.env.SETPAGE ? parseInt(process.env.SETPAGE) + 1 : 1;
-    const pageSize = process.env.ROWMEDIUM ? parseInt(process.env.ROWMEDIUM) : 30;
-    const field = 'proposalNumber';
 
-    const { loading, error, data } = useQuery(GET_PROPOSALS, {
-        variables: { pageSize, page, field }
-    });
+    if (title.proposalTitle.length > 0) {
+        return (
+            <div>
+                <Grid container className={classes.container}>
+                    <Typography variant="body1" className={classes.proposalTitle}>
+                        Proposals
+                    </Typography>
+                    {title?.proposalTitle?.map((row: any, index: number) => {
+                        return (
+                            <Grid item xs={12} className={classes.proposalCard} key={index}>
+                                <Card className={classes.card} elevation={0}>
+                                    <Grid container className={classes.container}>
+                                        <Grid item xs={8} sm={10} className={classes.proposalNum}>
+                                            <NavLink
+                                                href={`/proposal/${row.proposalNumber}`}
+                                                name={
+                                                    <Typography
+                                                        variant="body2"
+                                                        className={classes.value}>
+                                                        #{row.proposalNumber}
+                                                    </Typography>
+                                                }
+                                                textSecondary
+                                            />
 
-    const getProposer = (proposalNumber: number) => {
-        let proposer = '';
-        if (data && data.proposals && data.proposals.proposals) {
-            for (const d in data.proposals.proposals) {
-                if (proposalNumber === data.proposals.proposals[d].proposalNumber) {
-                    proposer = data.proposals.proposals[d].returnValues.proposer;
-                }
-            }
-        }
-
-        return proposer;
-    };
-
-    const findProposalStatus = (proposalNumber: number) => {
-        if (data && data.proposals && data.proposals.proposals) {
-            for (const c in data.proposals.proposals) {
-                if (proposalNumber === data.proposals.proposals[c].proposalNumber) {
-                    if (data.proposals.proposals[c].removed === true) {
-                        return <Chips actionResult="Removed" />;
-                    }
-
-                    if (data.proposals.proposals[c].status === 'Approved') {
-                        return <Chips actionResult="Passed" />;
-                    }
-
-                    if (data.proposals.proposals[c].status === 'Rejected') {
-                        return <Chips actionResult="Rejected" />;
-                    }
-                    if (data.proposals.proposals[c].status === 'Vote') {
-                        return <Chips actionResult="Vote" />;
-                    }
-                    if (data.proposals.proposals[c].status === 'Deposit') {
-                        return <Chips actionResult="Deposit" />;
-                    }
-                    if (data.proposals.proposals[c].status === 'Create') {
-                        return <Chips actionResult="Create" />;
-                    }
-                }
-            }
-        }
-    };
-    if (loading) return <ComponentLoader />;
-    if (error) return <ErrorMessage message={error.message} />;
-
-    return (
-        <div>
-            <Grid container className={classes.container}>
-                <Typography variant="body1" className={classes.proposalTitle}>
-                    Proposals
-                </Typography>
-                {title
-                    ? title
-                          .slice(1)
-                          .reverse()
-                          .map((row: any, index: number) => {
-                              return (
-                                  <Grid item xs={12} className={classes.proposalCard} key={index}>
-                                      <Card className={classes.card} elevation={0}>
-                                          <Grid container className={classes.container}>
-                                              <Grid
-                                                  item
-                                                  xs={8}
-                                                  sm={10}
-                                                  className={classes.proposalNum}>
-                                                  <NavLink
-                                                      href={`/proposal/${row.proposalNumber}`}
-                                                      name={
-                                                          <Typography
-                                                              variant="body2"
-                                                              className={classes.value}>
-                                                              #{row.proposalNumber}
-                                                          </Typography>
-                                                      }
-                                                      textSecondary
-                                                  />
-
-                                                  <Typography
-                                                      variant="body2"
-                                                      className={classes.proposer}>
-                                                      Proposer
-                                                      <NavLink
-                                                          href={`/proposal/${getProposer(
-                                                              row.proposalNumber
-                                                          )}`}
-                                                          name={
-                                                              <MiddleEllipsis
-                                                                  text={getProposer(
-                                                                      row.proposalNumber
-                                                                  )}
-                                                              />
-                                                          }
-                                                          className={classes.proposerAddress}
-                                                      />
-                                                  </Typography>
-                                              </Grid>
-                                              <Grid
-                                                  item
-                                                  xs={3}
-                                                  sm={2}
-                                                  className={classes.proposalButton}>
-                                                  {findProposalStatus(row.proposalNumber)}
-                                              </Grid>
-                                              <Grid
-                                                  item
-                                                  xs={11}
-                                                  sm={8}
-                                                  className={classes.proposalDescription}>
-                                                  <NavLink
-                                                      href={`/proposal/${row.proposalNumber}`}
-                                                      name={row.proposalTitle}
-                                                      textSecondary
-                                                  />
-                                              </Grid>
-                                          </Grid>
-                                      </Card>
-                                  </Grid>
-                              );
-                          })
-                    : null}
-            </Grid>
-        </div>
-    );
+                                            <Typography
+                                                variant="body2"
+                                                className={classes.proposer}>
+                                                Proposer
+                                                {row?.proposer ? (
+                                                    <NavLink
+                                                        href={`/account/${row?.proposer}`}
+                                                        name={
+                                                            <MiddleEllipsis text={row?.proposer} />
+                                                        }
+                                                        className={classes.proposerAddress}
+                                                    />
+                                                ) : (
+                                                    <NotAvailable variant="body2" />
+                                                )}
+                                            </Typography>
+                                        </Grid>
+                                        <Grid item xs={3} sm={2} className={classes.proposalButton}>
+                                            {row.proposalStatus ? (
+                                                <Chips actionResult={`${row.proposalStatus}`} />
+                                            ) : row.proposalStage ? (
+                                                <Chips actionResult={`${row.proposalStage}`} />
+                                            ) : null}
+                                        </Grid>
+                                        <Grid
+                                            item
+                                            xs={11}
+                                            sm={8}
+                                            className={classes.proposalDescription}>
+                                            <NavLink
+                                                href={`/proposal/${row.proposalNumber}`}
+                                                name={row.proposalTitle}
+                                                textSecondary
+                                            />
+                                        </Grid>
+                                    </Grid>
+                                </Card>
+                            </Grid>
+                        );
+                    })}
+                </Grid>
+            </div>
+        );
+    } else {
+        return <ComponentLoader />;
+    }
 };
 
 export default ProposalList;